refactor(desktop): convert PresetEditorSheet to a centered Dialog
Renames PresetEditorSheet -> PresetEditorDialog and switches the side
sheet to a max-w-2xl modal Dialog matching 1Code's MCP/agent edit
patterns. The body now uses a 2-column grid for short fields (Name +
Description) and full-width sections for Commands, Applies-to,
Directory, Launch mode, and a bordered Auto-run group.

Drops the custom field/control className overrides, the uppercase
"Advanced" subheading, the question-mark tooltips, and the
LabelWithTooltip helper. Fields use a small Field helper with a
hint underneath instead.
