--- apps/desktop/src/renderer/routes/_authenticated/settings/terminal/components/TerminalSettings/components/PresetsSection/components/LabelWithTooltip/index.ts
@@ -1,1 +1,0 @@
-export { LabelWithTooltip } from "./LabelWithTooltip";
