+++ apps/desktop/src/renderer/routes/_authenticated/settings/terminal/components/TerminalSettings/components/PresetsSection/components/PresetEditorDialog/index.ts
@@ -1,0 +1,1 @@
+export { PresetEditorDialog } from "./PresetEditorDialog";
