--- apps/desktop/src/renderer/routes/_authenticated/settings/terminal/components/TerminalSettings/components/PresetsSection/components/PresetEditorSheet/index.ts
@@ -1,1 +1,0 @@
-export { PresetEditorSheet } from "./PresetEditorSheet";
